refactor: split agent profile pane actions
